Add image drawing capability and make image view element
Also split the renderer into multiple parts (partial class) and small refactors.
