+++ src/CatUI.Elements/Themes/ImageViewThemeData.cs
@@ -1,0 +1,13 @@
+namespace CatUI.Elements.Themes
+{
+    public class ImageViewThemeData : ElementThemeData
+    {
+        public ImageViewThemeData()
+        {
+        }
+
+        public ImageViewThemeData(string forState) : base(forState)
+        {
+        }
+    }
+}
